Fix problem that caused subsequent tests run by runtests.py to fail.
In order to import jdb and other "lib" modules, test_jmxml.py
replaced the first item (".") in sys.path with "../lib".  This
works fine when test_jmxml is run as a standalone program but
when it is imported (by runtests.py) that change to sys.path is
persistent and affects other modules.  Corrected immediate problem
but full fix in to not make changes to sys.path in importable
modules.  Opened IS-133.