jmcgi.py: produce hist records in reverse order for web page.
Formerly the history records created for use in displaying
entries in the entr.py web page were produced in a old-to-new
order placing the most recent and usually most relevant at
the bottom of a sometimes very long display.  This update
simply revereses that order so the most recent ones appear
at the top.