jinja: convert updates.tal to jinja
Also partially fixed problem pre-existing in ealier TAL version:
updates.py rendered template with 'this_page' set to "entr.py".
The result was that logging in or out on the updates page returned
to entr.py instead of updates.py and the former responded with an
error because no entry number was given in url.  Fixed so that now
it will return to the updates index page.  Full fix should also
pass the current url params.