*.jinja: Revise page title format.
In layout.jinja (common base for all pages) swap the size of the
two displayed headers: deemphasize the general "jmdictDB" header
which is same on all pages from H2 to H3, emphasize the per-page
header from H3 to H2.
On all pages remove "JMdictDB" from the per-page header since it's
redundant (it's mentioned in the general header immediately above).
Use "JMdictDB" + the per-page header as the HEAD <TITLE> element.