cgi post methods, dbg parameter propagation
Remove 'method' from the context passed to the jinja templates; it was
always set in the cgi scripts dependent on "dbg".  We pass only dbg to
the jinja templates and they determine the form method appropriately.
Forms "edform" and "edconf" use the "post" method by default, but switch
to "get" if dbg is true.  The rest use "get" always.  Formerly some used
"post" to avoid putting the sid in the url parameters when cookies were
disabled but that seems pointless give that sid are also in links.
'dbg' is passed as hidden input in all forms (but not in links).

jmcgi.py also contains the sid fix from hg-2018043-6b7b25.