logger.py: fix fail to write non-ascii message to logfile
When Python logging module used in a cgi script (at least in the case of
Apache-2.4, not sure if this is server dependent) to write mesasges to a
logfile, a message with non-ascii characters results in a UnicodeEncodeError
exception which is handled by the logging module code by writing the error
to stderr (causing it to end up in the webserver's error log file).  The
orginal message is dropped and never appears in the given logfile.
The workaround seems to be to explicitly open the log file with utf-8
encoding so the UnicodeEncodeError doesn't happen.

Also, check if the log file is writable; if not writable (because it
doesn't exist, permissions, etc) write a log message to stderr and
disable further logging.  The intent is to provide notice that the
logging file is inaccessible without generating such messages on every
attempt to wtite a logging message (although they will be generated
for every cgi script excecution which might be worse.)