Fix: JEL syntax error message lost formatting
Error page message for JEL syntax error was being displayed as regular
html text rather than formatted with "pre" style causing everything to
appear on one long line and the carat marking the error position to
be completely mislocated.  The need for "pre" for JEL errors was
overlooked during the conversion from TAL to Jinja templates.

The current fix is interim; the sense text appears as one long line
as it did before but looks worse now because it extends beyond the
borders of the "green box" (previously it appeared on a white page
with no visible borders so it didn't look so bad.)