JELparser: Improve parse error message by preserving newlines.
Change the character used to "glue together" the kanji, reading and
sense text when forming the input for the JEL pasrser, from a newline
character to form-feed character.  This allows newline characters
in the sense text to be preserved (but semantically ignored) preserving
the input format used by the user when showing in an error message.

Formerly newlines were stripped from the sense text with the result
that an error message would show the sense text as one long line.