Fix: nl and cr characters getting into gloss text.
Some recent submissions at EDRDG have had newline and cr characters in
the gloss text strings saved to the database.  This started occuring
after the change (hg-20180523-6b1a12) to leave newlines in the submission
text in order to improve error messages.  That commit removed a function
call in edconf.py that translated and compressed all whitespace in the
sense box text into spaces.  That should have been ok because the JEL
lexer/parser also does whitespace compression in gloss strings when
building an Entr object.  However it appears that the JEL code was not
working correctly (perhaps that's why the compression was added in
edconf.py?)  This commit fixes the whitespace compression in the JEL
lexer/parser.