*.sql: Make execution independent of current directory
postload.sql and reload.sql both formerly used the psql "\i"
directive to run another .sql file as a sub-task but because
the filename to run is relative to the current directory, the
caller had to be cd'd correctly for the .sql file to be found.
Replace "\i" with "\ir" which locates the .sql file relative
to the location of the calling file and is thus independent
of the caller's current diretory.
"\ir" has been available since Postgresql-9.2.