xresolv.py: if no upserts, exclude ON CONFLICT clause from INSERT
If upserts not requested, create the INSERT sql without the
  ON CONFLICT clause.  This was supposed to restore the execution
  time back to ~10m but failed to (perhaps because of addition of
  the deletes mentioned next?  Seems unlikely but...)
Reimplement (not)--keep option as a delete of the corresponding
  xresolv row in wrxref() after the xref is successfully written.
Restore processing in groups of one entry.  This was case in
  hg-180619-55b627 but was temporarily lost in change to use
  a query to retrieve all target candidates (hg-180630-7e0acf).
  We don't currently do anything (eg save or rollback) at entr
  group level.
Add --start, --stop options.
Note: handling of errors and implementation of partial rollbacks
  is still incomplete.
Note: execution time for examples->jmdict is still ~50m.