Makefile: don't install config.ini, clarify in README.txt
Formerly 'make web' would install the config.ini file in the
development tree (in python/lib/) into the production lib
directory.  This was problematic because the dev version is
used when the app is served directly from the development tree
and one might want a different config.ini for that than used
in the production app.  Makefile 'web' target no longer installs
config.ini; the user is responsible for configuring and manually
installing config.ini in both the production and development
locations.

Updated README.txt to clarify the above and make clearer the
distinction between the two locations.  Also typos and other
minor edits.