entr.jinja: disable notes about asterisk-marked xrefs.
As mentioned in the template code, the notes about asterisk-marked
xrefs really should be presented only if there actually are xrefs
with the noted characteristics, and should appear closer to them,
not at the bottom of the entry.  Will postpone displaying the notes
(the asterisks are still applied to the xrefs when the given conditions
apply) until those issues can be addressed.