Add redirection replacement for command line conj.py demo program
python/conj.py was a command line program demonstating the use
of the JMdictDB conjugation tables in a stand-alone, non-JMdictDB
program.  It was extracted into an independent repository and
removed from JMdictDB in jm:180610-1f524c. This commit restores
a file of the same name that contains info pointing to the new
location.