Add lib/db.py
Add lib/db.py and add to LIB files in Makefile; this file is imported
by lib/xresolv.py but was overlooked when that change was made.  Also
imported by srchres.py in recent commit to get access to psycopg2
exceptions.  Intent is to use in future for basic database operations.