Replace jdb.get_query_cost() with postgresql statement_timeout
Previously cgi/srchres.py called jdb.get_query_cost() to estimate
how long a query would run.  The latter extracted the "cost" value
from the results of running an Postgresql "analyze" statement.
This was not always accurate nor robust.  Instead we now implement
a new "timeout" parameter in jdb.dbread() that will, if the executing
query takes longer, cancel the query.  This should provide a much
stronger limit on excessive resource consumption which is needed
with the addition of the regex search feature (jm:190227-034039)
to reduce the possibility of a regex denial of service (ref: redos
in wikipedia although postgresql's regex engine is already immune
to the specific examples given there).  If not overridden in the
config.ini file, a default timeout of 3000 mS is used in srchres.py.

srchform.py: fix typo in form text.