Strip '\r's from history comments and refs at submission
In the jmdictdb instance at edrdg.org, it was noticed that many history
comment and refs texts appear to be wrapped to fit in 80 columns even
when that broke long strings like urls and additionally the wrapping
was done with '\r\n' end-of-line characters rather than canonical
'\n's.  I've not been able to reproduce in a Linux/Firefox environment,
suspect MS Windows may be involved.  This update removes '\r' characters
from the comments and refs text before writing to the database.  It
does not address the wrapping issue as that may be a client-side issue.