xresolv.py: fix crash due to missing view "rslv".
Running xresolv.py during a test load of jmdict.xml resulted in a
failure in python/xresolv.py when it tried to run query based on
non-existant view "rslv".  Added the missing view to pg/mkviews.sql
based on the comments at bottom of resolv.py that gave its definition.
View was introduced in commit 180630-dd6fea9 but was apparently never
added to the db schema.

Running again resulted in crash due to db check constraint violation:
  IntegrityError: new row for relation "xref" violates check constraint "xref_check"
Problem was that view rslv, which finds candidate xref target entries
during the resolution process, was including in the candidates list
under some circumstances, the entry whose xref was being resolved.
Specifically, entry 2817960 (応え) has a see=応え xref with the intent
of referencing 1449530 (答え;応え).  But view rslv included entry 2817960,
the referencing entry, in its results as well.
Modified rslv in mkviews.sql to exclude the referencing entry.

Also:
- Renamed view "rslv" to "vrslv".
- Extracted some of the unused views in mkviews.sql to new file
  supl-views.sql.
- Removed the comments with the erroneous rslv view definition from
  xresolv.py
- Some cosmetic adjustments in mkviews.sql.
- Untabified and stripped trailing whitespace from mkviews.sql,
  xresolv.py (this accounts for all the changes shown in the diffs
  that look unchanged.  Use a diff option that ignores whitespace
  to see the significant changes.)