jmxml: fix gloss g_type ignored, remove jmparse -y option
Prior to this fix, "g_type" attributes in glosses were erroneously
ignored when parsing an xml file with jmparse.py.

Also, remove the jmparse.py --extract (aka -y) option and supporting
code (xlit parameter) in jmxml.py that would convert the text "lit: "
in a gloss into a ginf "lit" value.  This was an early hack when literal
glosses were indicated that way in the EDRDG JMdict XML file but there
are no longer any entries where it is still used.