cgi: move logger.enable() call into main()
Previously in all the cgi/* scripts logger.enable() was called in
module scope:
  import logger; from logger import L; logger.enable()
logger.enable() replaces python's top-level exception handler with
one that will produce an html error page rather than the default
stack trace.  But this created a problem when any of these cgi
scripts were imported by some other program (in order to get access
to internal functions not available in lib modules or to run unit
tests on internal functions) because the logger.enable() statement
would be executed as a side effect of the import, replacing the
importing programs's exception handler.  When an error occured the
program would then print the html from the jmdictdb error page
instead of a stack trace. :-(

All cgi/* scripts were updated to move the logger.enable() call
to inside the script's main() function to prevent the above side-
effect.