ylex.py, yparse.py: update to Ply-3.11 base
These files are the lexer and parser for parsing the YACC syntax
used in jelparse.y.  They are modifications of the files of the
same names shipped with Ply (third party parser-generator).  The
modifications allow them to process the Python code snippets used
in the jelparse.y action rules rather the the C code actions they
were originally written for.

This revision migrates those local changes from the old versions
which were based on old, unknown versions of Ply (likely in the
version 2.2 to 2.5 range) to the current Ply version 3.11 (although
as it turns out there were very few non-whitespace upstream changes
in that time).