jdb.py: remove obsolete getattr() calls
In jdb.addentr() and jdb.setkeys() there were a number of unneeded
calls to getattr().  These were a remnant from when jmdictdb objects
omitted attributes that had empty values.  For many years now all
objects are created with every attribute present so there is no
longer a need to wrap attribute accesses with getattr() to protect
in case the attribute is not present.