jdb.py: entrList() returned Entr's without '.chr', '._krslv' attributes
For a long time policy has been for Entr and its component objects
to always have a full set of attributes to avoid the need to wrap
every attribute access in a getattr() call.  However, the Entr
objects returned by jdb.entrList() did not have '._krslv' or '.chr'
attributes if the corresponding data was not present in the database
entry(s).  Problem was in jdb.mup() which skipped assigning a
'pattr' attribute if there was no matching parent object.
One side effect of the fix: jdb.mup() is also used by jdb.augment_xrefs()
to add a .TARGET attribute to xrefs.  Before, if no target was found
no .TARGET attribute was added.  Now, a .TARGET attribute is added with
a value of None.  I did not see any use sites where this would matter
and all tests still pass.

jdb.add_hist(): formerly returned a Obj item, now more properly returns
a Hist object.

entr_bld(): The repr of an Entr object read from the database would
print the _rdng,_kanj attributes in that order while an Entr object
created by direct instantiation would print the in the order _kanj,_rdng.
This was a cosmetic difference since access is always by key but it is
easier to reading debugging output if the orders match.  The order of
mup('_rdng',...) and mup('_kanj,...) calls in entr_bld() resulted in
the former order and were swapped to produce the latter order.

jdb.entr_data(), jdb.entr_bld(): minor code rearrangements and comment
revisions.