Move classes DbRow and Obj from object.py to db.py
There were separate, divergent, implementations of DbRow in both
objects.py and db.py.  The former was used as a superclass for
all the JMdictDB objects (Entr, Rdng, etc) and were returned by
the functions that access the database via a jdb.dbOpen() cursor
(jdb.entrList(), jmcgi.get_entrs(), etc).  The db.py version was
to demo the ability to return DbRow objects from a Psycopg2
cursor using psycopg2's standard DBAPI interface.  The intent is
to move to the latter interface throughout the JMdictDB code in
the future.

This revision modifies the DbRow class in db.py so that it will
work as a replacement for the objects.py one, and makes it the
canoniocal version, removing the objects.py version.  Its super-
class, Obj, was also moved from objects.py to db.py, as were a
couple supporting functions (_compare(), _p()).

All tests currently pass including the DbRow-specific ones.