Restrictions revamp
Formerly a "restr", "stagk" and "stagr" restriction was represented
internally as a (Restr, Stagr or Stagk) restriction object that was
referenced from a list on the primary object (eg Rdng._restr) and
also from the secondary object (eg Kanj._restr).  This was awkward
because to determine restrictions one had to scan both lists and
identify the restriction objects common to both.

This revision dispenses with the second list.  There is now a list
only on the primary (Rdng or Sens) object and the "other" (Kanj or
Rdng) object of a restriction pair is identified by an integer
number in the restriction object that gives either the base-1 position
in the Entr's ._kanj or ._rdng list or Kanj.kanj/Rdng.rdng number.
(Reasons for the dual interpretation are explained in restr.py.)

The revision functions were extracted from jdb.py into their own
module, restr.py.

The add_restr*() functions are no longer needed outside of restr.py
and have been replaced where they were used (mostly in tests).

Function headword() moved to the xref group of functions in jdb.py
since that is where it is used.  Tests were added from restr_extract()
which is used only by headword and verified they produce same results
before and after changes.

Permissions on python/xresolv.py were corrected to add "x".

All tests pass (although coverage of restrictions is scanty.)
Export of current EDRDG jmdict database to xml compares equal to
a jmdict.xml file downloaded from EDRDG (ie, one produced with
rev 190401-69c9857) sans a small number of changes due to current
edits or other explicable changes.
An import of a jmdict.xml file to database followed by export back
to xml shows a numbers of differences in xrefs (mostly that addition
of sens numbers to xrefs) that needs to be looked at.

Documentation (eg schema.odt, schema.dia) has not been updated yet.