xresolv.py: was failing to report "not found" targets
An error in the generated SQL used to get the set of candidate
targets was leaving out xresolv table entries for which there
were zero candidates.  This did not affect results (since those
items are unresolvable anyway) but did cause them not to be
reported in the logging messages.

Also changed permissions to make xresolv.py executable.