Tests: jmdb.py tests db manager now checks db hash
Prior, jmdb.py would simply check the filename a test database
was loaded from to determine whether or not it was the requested
one or whether it had to be reloaded.  However, it is likely that
changes will be made to tests databases in different banches
without a name change so a robuster check is needed.

jmdb.py now takes the hash of the source file and saves it in
the database; before running a test with a requested database,
it checks the hash of the requested database and if different
from the stored hash, reloads the database.