027-835781.sql: add missing views vrslv and vrkrestr (db-835731)
The view "vrslv" (and its helper, "vrkrestr") is used by xresolv.py
when resolving xrefs.  It was originally overlooked when xreolv.py
changed to use it but was added to mkviews.sql in git-190425-aed321f.
At that time it should have been added to a database patch file for
incremental upgrade but was again overlooked.  Including it in
patches/027-835781.sql since that patch has not been applied to
any operating JMdictDB instances yet.