IS-248 jelparse.y: exception: "nokanji" tag with entry with no kanji
When the JEL parser was given a reading with a [nokanji] tag but
the entry had no kanji, an IndexError exception was raised in
restr.txt2restrs().

It is non-sensical to apply a "nokanji" tag to a reading when
there are no kanji on the entry to actually exclude and a correct
response is to raise a jelparse.ParseError exception.  It appears
an attempt was made to check for this condition and do the right
thing but the check was erroneous.
Fixed the check and added test for proper response to this condition.