IS-249: New entry submission fails in empty jmnew database
A freshly created jmnew database (created by 'make jmnew') has a
kwsrc table entry for only a "test" (id=99) corpus.  Several places
in the cgi/edconf.py code check a new entry by verifying its .src
attribute is the same as that of the "jmdict" corpus.  The of course
fails when there is no jmdict corpus.

Since commit 140627-f0b0630d (IS-228) a corpus type number is
available and the edconf.py checks were changed to use that which
fixes the problem.