xrslv: jelparse.y: parse xrefs into unresolved refs
Prior, the JEL parser converted xrefs in JEL text (eg "[see=...]"
into an internal format, and then called an internal function that
would try to resolve them, creating Xref objects when successful
and Xrslv objects (for unresolved xrefs) when not.

This required the jelparse to have access to a populated JMdictDB
database for the xref resolution process, which was a distasteful
mixing of concerns.

In this revision, jelparse parses *all* JEL xrefs into unresolved
xrefs (Xrslv objects).  Note that this is also how jmparse.py parses
XML -- the unresolved xrefs are resolved later in a separate step,
independent of the parsing.

As part of this implementation, functions resolv_xrefs() and
find_xref() were removed from jelparse.y and a new function,
xresolv() that does the xref resolution they formerly did,
added to module jdb.  (Actually, for back compatibilty,
jelparse.resolv_xrefs() was turned into a thin shim function
that calls jdb.xresolv().

Note: the jelparse resolution functions that were removed did caching
of lookups and that functionality is not (yet?) in the replacement
jdb.xresolv() function.  Will need to see if this is a problem.

All tests currently pass.