tests/jmdb.py: reload test db on only file hash difference
Previously jmdb.DBmanager would force a reload of a test database if
the filename requested was different than the filename stored within
an existant loaded database.  But it is often the case that one will
run tests in a cloned copy of the software which will cause any test
database to be reloaded because their filenames are different even
though they are identical to what is already loaded based on the file
hash.  To avoid these unnecessary reloads, DBmanager now ignores
filename differences and forces a reload only if the file hash is
different.