jmapp: Advanced srchform migrated
The cgi code reshaped sets of checkboxes into blocks
in order to fit the width of the blocks to the expected
web page width.  This reshaping remains in the code in
rest.py for expediency but is inappropriate there since
the rest.py code should be view independent.  It was
marked with a FIXME for later attention.