jmjapp: Search results (migrate sql search page)
Added the sql search page (srchsql.jinja) and supporting code.
The code in the form template that checked that sql searches were
configured as available and the user was logged in as an editor
was removed; these checks are still done on submission but since
the page is not advertised for general use, there is no need for
a preemptive warning.

Add error page response for some search problems.
srvlib.login_handler(): fix error if validate_user() call
  returns None.