jmapp: edform migrated
Used a different strategy.  Previously the view in jmapp was responsible
for extracting the url parameters and would call a function in srvlib.py
(if the function needed access to Flask) or rest.py (if no Flask access
needed) that would do whatever work was needed and return data to the
jmapp view function which would render the page template with it.

For edform, the cgi code was moved (with minor adjustments for the
flask wnvironment rather than cgi) to a separate module in lib/views/
The code continues to do the url parameter extraction, and returns
the result data to the jmapp view for rendering as before.
This allows the cgi code to be reused with fewer changes.  Putting
it a separate file allows maintaining the preexisting large volume
of comments (in the edform case) or code without bulking up the shared
srvlib.py file.  The jmapp view now does nothing more than call the view
function and (as before) render the returned results.