jmapp: edconf migrated
In the cgi version the /edconf.py endpoint was accessed using a
POST request.  This has been changed to a GET request since it
is an idempotent request.  Another reason for the change was that
it was difficult to get the login form to work with a POST request
(an initial attempt resulted in breaking logins for all the
GET pages.)