fix: cgi user page error: DbRow has no attribute 'copy'.
When opening the cgi user.py page by clicking username when logged
in was getting an AttributeError: 'DbRow' object has no attribute 'copy'
This was happening in:
  File "/home/stuart/devel/jdb/jb/web/cgi/user.py", line 65, in sanitize_o
    o = obj.copy()
Problem was that the copy() method had been lost when moving the DbRow
class from objects.py to db.py in 190530-6dc0325.  Restoring the method
fixed the problem.