jmapp: empty form values returned as '' rather than None
Functions that retrieved form values from empty text boxes returned
them as an empty string ("") rather than None as desired.  In the
case of edconf.py this resulted in sending an Entr object with a
.seq value of "" to view/submit.py where it resulted in an exception.
For userupd.py, it resulted in changing user database values like
"email" and "fullname" from None to "".