jmapp: remove the "dbg" url parameter
The "dbg" url parameter was used in the cgi code mostly to force
forms that normally use the POST method, to use GET insead so that
the parameters could be copied from a browser's address bar and used
for debugging.  Flask allows debugging via its development server so
the dbg parameter is no longer very useful.