jmapp: support for maint and excessive load status
Removed contact name and email from the status_maint.html
and status_load.html pages because they need to be dynamically
generated from info in config.ini but contrarily we want to
use a static html page and not rely on app to generate.
(Although we *do* rely on app to decide  whether to show
the pages or not.)
Also changed the status_*.html pages to use html5.