Makefile: add modules restr.py, submit.py, dbver.py; file .htaccess
The lib modules should have been added in commits:
  190609-620303c (restr)
  190502-881bca2 (submit)
  190514-0f51e18 (dbver)
but were overlooked at the time.

The .htaccess file is to restrict access to the lib directory from
the web server which in important because lib contains some sensitive
information (eg config.ini).  Some additional configuration of the
Apache webserver may be needed to allow the .htaccess file to be
operative.