cgi: remove default db config settings from config.py
It is possible a site will have a Postgresql database configured
to use "trust" authentication in which case, any database user name
will be able to connect without needing a password.  To disallow
connections to a specific service (svc=...) in such a case it is
cleanest to simply remove the entire corresponding database section
from the jmdictdb config.ini file but that won't work if the section
is defined by default.  Hence the removal of all the database default
sections.