JEL (jelparse.y): Resolve xref corpus to id during parse
With the change in git-190827-8b86f59 to parse xrefs to Xrslv
(unresolved xref) items initially, we can no longer use '' as
a corpus value (which meant "resolve within the current corpus)
because the table "xresolv" expect an integer value or NULL.
Database errors were occurring when trying to write Xrslv objects
to that table when the corpus field (.vsrc) contained a ''.

Consequently the convention for "current corpus" was changed
from '' to None, and the convention of None meaning "resolve
across all corpora" was dropped.  I don't think the ability to
resolve across all corpora is needed because:
- There was no JEL syntax to produce a None corpus value.
- Resolution across all corpora is not necessary since in most
  cases a submitter will have a specific entry in mind to
  reference and will know (or can easily determine) the corpus
  in which it resides.
- A search across all corpora is likely to have a high failure
  rate due to multiple matches.

(It might have been better to keep the xref corpus names as
strings and change the database "xresolv" table to use TEXT for
the corresponding column ("vsrc"); this would allow accepting
xrefs for corpora that will be loaded later.  The tradeoff is
typos won't be caught immediately.)

Also changed xrefs to entry id numbers to be distinguished from
xrefs to seq numbers by using a negative value for id numbers.
This is incomplete: none of the receiving code knows about this
convention yet.