xresolv.py: exclude non-active entries from xref resolution
When running the xresolv.py program on the current edrdg jmdict
database it crashes with a duplicate primary key error.  Problem
was traced to a number of rejected entries that contained both an
unresolved and a resolved xref to the same target entry.  When
xresolv.py resolved the unresolved one, the result was a duplicate
of the preexisting resolved one.

This rev excludes non-active (ie rejected and deleted) entries
from the resolution process allowing xresolv.py to complete
successfully.  The source of the xref pairs is still unknown.