Tests: jmdb.py: better error handling when db open fails
Raise an exception if an unexpected error occurs when opening a test
database rather than treating in the same way as expected errors
(bad hash value, non-existant database).  Since such errors are
usually persistent and the database is often opened in a test module
setup function, raising will result in the abortion of all module
tests which is more desireable than repeating the failure in each
test or test class.