JEL: support for an "extended jel" syntax
JEL was originally intended for only the kanji, reading and
sense parts of an entry.  However it is a succinct way to
describe entries for other purposes such as test data.  (XML
is too verbose, Edict format has ambiguities and can't be
changed for back-compatibility reasons, JSON is not as readable.)
This revision is an attempt to extend the JEL syntax to allow a
fuller specification of entries by providing for:
- multiple entries in a stream
- comments
- entry description on a single or multiple lines
- additional entry metadata: id, seq, corpus, stat, unap, dfrm
It does not yet provide for specifying some entry information,
in particular entr.srcnotes, entr.note, groups, kanjidic related
info.

jel.py: new module with functions for parsing "extended JEL"
  syntax.
jelload.py: new program that uses jel.py and will load entries
  into a database from an extended JEL text file.
fmtjel.py: changed the format of the entry header slightly to
  match the format accepted by jel.parse().  Example:
    previous: jmdict 1000010 A* {1685508}
    new:      1000010 jmdict A* [id=1685508; dfrm=120778]
  The use of square brackets and semicolons (replacing curly
  brackets and commas) is to make the syntax more ammenable to
  parsing directly in the yacc JEL parser in the future since it
  already lexes those character specially.
  Also a few code cleanups (eg replace iif() function with
  ternary "if" statements.)
test_fmtjel.py: update for header change in fmtjel.py..