is251: srchform cgi page crash when new kwfld item added
The srchform.py page was crashing when an additional row
was added to the database "kwfld" table.  Problem was that
the additional item created a new html row of field items
with all but the first item set to None (as expected) but
the jinja template was not prepated for a field item with
a value of None.  Modified the template to accept None
values.