Change MISC "lit" tag to "litf", was confused with GINF tag.
There is already a GINF (gloss) "lit" ("literaly") tag and when
an entry with that tag was edited, the GINF tag was removed and
a sense MISC "lit" tag added (since the "[lit]" in the JEL text
is parsed with higher priority for sense tags.  Note that one
can explicitly disabiguate the GINF usage using the JEL syntax
"[ginf=lit]" but it is easier and less confusing to simply give
the MISC tag a different name.

This update made for existing database version db-7256aa.