entrs2xml.py: fix unwanted buffering of progress dots
Program prints "." characters to stderr to show progress.  Sometime
ago it stopped doing this and up to now would print them all out at
once at the end of the program.  A Stackoverflow answer seemed to
say that stderr changed from unbuffered to buffered with Python-3
but I'm surprised I've not noticed it before.
This rev restores the desired behavior.