entrs2xml.py: Remove multiple corpus capability
Formerly the --corpora (-s) option of entrs2xml.py accepted
multiple corpora.  This was originally intended to allow dumping
an entire database of multiple corpora to XML using the extended
dtd.  However, that ability has not been used, costed code
complexity and has an alternative: dump corpora one at a time,
all but the first without a dtd, and stitch the the files
together manually.

Other changes:
- Remove --encoding option; encoding is now always utf-8.
- Replace the --database, --host, --user, --port options with
  single --dburi option.

Compared outputs of old and new versions of program on jmdict
and jmnedict corpora in "compat" mode and got no differences.
Did not check the extended output.