entrs2xml.py: major code reorganization
Processing changed to separate selection of entries vs retrieving
blocks of them.  Instead of constructing a single sql statement
to retrieve blocks of entries run in a loop, we construct a base
statement that selects all entries of interest, then use that as
a subselect in the sql that retrieves blocks.  The latter step is
now done in a generator function that yields blocks of entries.
Performance between old and new versions was checked with no
significant differerences seen.

Code was majorly refactored to simplify and clarify by extracting
much code formerly in main() into functions.