entrs2xml.py: require -o for output file
Also allow sequence numbers as command line arguments.
Formerly the output file was specified as a command line argument.
We now use command line arguments for sequence numbers and the output
file is given with the -o (--output) option.  As before, if no
output file is given output is to stdout.

Fixed errors:
- NameError for variable 'seqfile' in setup_progbar().  This occured
  when seqnum arguments or the --seqfile option was used.
- "jmnedict" not accepted as valid --compat value.

Kmown errors:
- AttributeError exception in write_entrs() when no --compat option
  given.