entrs2xml.py: --compat option no longer required (usually)
Program now derives the right value of --compat from the
--corpus value so usually only the latter option need be
given on the command line.  --compat can be given to select
one of the legacy optionsm or (in conjunction with new
--force option) to use an incompatable --compat option
(eg to format jmdict entries as jmnedict xml).

As a consequence of the above --corpus/-s is now a required
option.

Compared outputs of old (pre-recent changes: 190508-02d9fdb)
and current versions of program on jmdict and jmnedict corpora
and got no differences; also jmdict in jmex compat mode, also
no difference.