entrs2xml.py: correct help text re output ordering
The help text formerly said order of entries in output XML were
in the same order sequence numbers were given on command line or
in seqfile.  That was wrong, entries in XML are always written in
sequence number order.

Also expand description text to give a quick intro to common
options and do minor tweaks to texts of other options.

seqnum arguments and --seqfile option tested.