hggit.py: add --long option, better --dir default
Changed default value of --dir option to location where the hg2git
mapping files are actually kept (doc/hg2git/) so it doesn't need to
be refound every time I use the program.
Added a --long option that prints what was printed by default before:
both the supplied hash and the found equivalent.  The default id now
a "short" output consisting of the found hash type ("git" or "hg"),
a dash and the hash value.