Tests: move JelParser class into jmdb.py
Because JEL is a concise and conventient way to construct Entr
objects for tests, a JelParser class was duplicated in several
test files.  They have been consolidated moved into a common
class in jmdb.py.

The JelParser class in jmdb.py contains some fixes not in the
original test files.  In particular the handling of the .parse()
method's 'unap' argument was corrected.  A new 'dfrm' argument
was also added.  Added ability (enabled by default) to augment
any xrefs in the entr.

Also removed the __unittetst=1 lines from test_jel_restr.py and
test_jel_xrefs.py.  Don't see anything about this in unittest docs
and think it is an artifact of times gone by.