test_jdb_kwds.py: remove kw tests, add quote tests
Formerly test_jdb_kwds.py contained a list of keywords (aka
tags) which were compared to the tags in the pg/data/*.csv
files.  When adding or updating tags, changes needed to made
to both the *.csv files and the list in test_jdb_kwds.py.

Since tag changes also require the creation of a database
update file to apply to operational databases and that update
is always applied to the jmtest01 database, we can eliminate
the need to check the csv files against a duplicate tag list
by checking against the tags against the jmtest01 database.

We now use a new test module, test_kwds.py to do that check
and the duplicate tag list and associated tests have been
removed from test_jdb_kwds.py (which really should be testing
only the correctness of the jdb.Kwds class and not specific
data contents.)

Also add tests for descr field with quotes.