Remove last traces of 'gA' freq scale
The "gA" frequency scale was for a set of Google "hit" counts
for various words in the jmdict corpus similar in concept to
the "nf" 1-48 scale and like it, defined in the kwfreq tags.
The data for it was collected by Kale Stutzman in 2007 but was
never part of the JMdict XML distribution, it had to be added
to the database separately.  Today, Google hit counts are
meaningless.  The "gA" options were removed from the web UI
some time ago and this revision removes (hopefully) the last
traces of them in the code and elsewhere.