pkg: Make jmdictdb an installable python package
This is intended to allow installation of a stable release of the
JMdictDB code which will be used by scripts, the web server etc,
while allowing a developer on the same machine to work on and test
a development version without impacting the installed version.

The code formerly in ./python/lib was moved to ./jmdictdb, made a
python package and a setup.py script added to allow installation
as a system-wide, importable, "jmdictdb" package.  The executable
programs in ./python were moved into ./bin/.  Tests were moved out
of ./python to a top-level directory.  Although not necessary for
the packaging goal, took this opportunity to rename ./pg to ./db
and rename the former ./patches directory to ./db/updates.  This
directory restructuring was done in a previous commit.  This
commit adjusts the import statements in the library (jmdictdb
package), an installable script (./bin/shentr.py) and three
cgi scripts (srchform.py, srchres.py and entr.py in ./web/cgi)
to verify the approach.

Because of the location changes, all the library, script and
cgi code required changes to their import statements.  sys.path
manipulation is used so that if there is a jmdictdb package
in the location provided by the distribution (ie in ../jmdictdb
relative to scripts or ../../jmdictdb relative to cgi scripts, etc)
it will be used on the assumption that the code is being run in a
development environment.  If not, that is a program, cgi script,
etc is someplace like /usr/local/bin/, the assumption is that it
has been installed outside of a distribution tree and it will
import and use the installed jmdictdb library code.

We keep the installed lib/ directory but it now contains only
the config files (and a .htaccess file to prevent reading from
web).

./Makefile was revised to install the jmdictdb package.

Removed BOM from jmdictdb/edfmt.py, tests/tests/test_edconf_realign.py.
.gitignore updated.

At this point the Makefile can install the jmdictdb package, command
lines tools and cgi files.  For cgi, the srchform, srchres and entr
pages work, the shentr.py command works (and least generally, it
hasn't been updated in years.)

Changes in this revision are incomplete: modules import statement
were revised to allow them to be imported without errors, three
cgi scripts (the srchform, srchres and entr pages) and one command
line program (shentr.py) were updated to work with the new layout
in order to demonstrate feasibility.  The Makefile was updated to
install the software, either system-wide if run as root, or per-
user otherwise.  The other cgi, program and tools scripts and tests
will be converted later.  None of the Makefile's database targets
are expected to work.