pkg: fix user install; install only changed files
Previously, the cgi files installed per-user (in ~/public_html/
by default) were unable to import the per-user install jmdictdb
package (in ~/.local/lib/... by default).  This was not a problem
for command line programs because Python adds the per-user package
directory to sys.path automatically when the user runs them but
this does not happen for the cgi scripts that are run by a web
server owned by a different user.  The fix was to have the cgi
import a module (libpath.py) in the cgi directory (which is on
sys.path automatically) that adds the jmdictdb package directory
to sys.path.  The module is created by the makefile during install
which it can do so because it knows where the the jmdictdb package
is going.  Similarly when accessing the cgi files directly from a
development directory the libpath.py will have a different value
in it that gives access to the jmdictdb package in the development
tree.

When installing files to their target install locations, using
'make's "replace if source file is newer" policy is undesirable.
When checking out different branches from VCS it is easily possible
to get changed files that are older that the installed versions
and would wrongly not be replaced by 'make'.  We add a new script,
tools/install.py, that accepts a small subset of the system 'install'
program's options.  It compares target installed file and source
file and only installs the source file if it has different content
or mode.  This it not for efficiency (reinstalling all the files
every time 'make' runs is plenty fast enough) but to provide feed-
back to the developer that the files installed were the ones he/she
expected to be.

setup.py: add the dtd-* and *.xsl files to package-data.