pkg: split Makefile, separate targets for system/user install
The database related targets (jmnew, loadjm, etc) and software
install targets that were in Makefile are completely independent
so extracted the database targets into new Makefile.db.  Running
a database target now requires a command like:
  make -fMakefile.db loadall

Heavily revised Makefile which now has only targets related
to installing the software.  Instead of providing a single
"install" target and doing a system or user install based on
whether the user is root or ordinary, there are now two targets:
"install-sys" and "install-user".  Each has a check that allows
only the appropriate user to run it (root for "install-sys",
non-root for "install-user").  This will hopefully catch doing
an install as the wrong user in a more helpful way.

Provided a full set of uninstall targets.