Makefile.db, data/: create data/ directory when needed
For currently unknown reasons when there is a lot (GBs) of files
in data/, Python's 'pip' take an inordinately long time to do an
install.  This doesn't happen if data/ is a symlink to the same
files in a directory located outside jmdictdb project tree.  data/
was removed from the git repo and Makefile.db adjusted to create
it if it doesn't already exist, allowing a symlink to be used.

Some comments in Makefile.db were prefixed with tab characters
causing them to be treated as shell commands and echoed when the
containing action was run which was inappropriate.  Indented with
spaces instead so they are treated as makefile comments and not
echoed.

setup.py: replaced call to find_packages() with explicit list
of (one) package(s) in attempt to resolve slow install problem
mentioned above.  Had no effect but reduces complexity so leaving
it in place.

config-pvt-sample.ini: correction to comment.